Add shutter-speed and iso value to get_parameters so Focal finds them.
This patch adds "shutter-speed" and "iso-value" to the get parameters.
It is needed for ISO and shutter speed setting.

Also it fixes that shutter speed cannot be reset to "auto".

PS2: Added prefix ISO to all iso-values to conform with focals
requirements. As I understand this should be the new standard for
camera wrappers. See comments at:

http://review.cyanogenmod.org/#/c/48737/

Change-Id: I7151a79fa72cd4bb4229117160156fcffe349aef